third_party/blink/web_tests/platform/mac-mac10.15/external/wpt/preload/reflected-as-value-expected.txt

This is a testharness.js-based test.
[FAIL] Link preload "as" value for "video" should be "video".
  assert_equals: expected "video" but got ""
[FAIL] Link preload "as" value for "audio" should be "audio".
  assert_equals: expected "audio" but got ""
Harness: the test ran to completion.
